refactor: extract ReleaseSection as a component from snapshot page
This makes it easy to test this component
